add assessment method: hyperiqa
