--- facexlib/assessment/targetnet.py
@@ -1,62 +1,0 @@
-import torch.nn as nn
-from torch.nn import functional as F
-
-
-class TargetNet(nn.Module):
-    """
-    Target network for quality prediction.
-    """
-
-    def __init__(self, paras):
-        super(TargetNet, self).__init__()
-        self.l1 = nn.Sequential(
-            TargetFC(paras['target_fc1w'], paras['target_fc1b']),
-            nn.Sigmoid(),
-        )
-        self.l2 = nn.Sequential(
-            TargetFC(paras['target_fc2w'], paras['target_fc2b']),
-            nn.Sigmoid(),
-        )
-
-        self.l3 = nn.Sequential(
-            TargetFC(paras['target_fc3w'], paras['target_fc3b']),
-            nn.Sigmoid(),
-        )
-
-        self.l4 = nn.Sequential(
-            TargetFC(paras['target_fc4w'], paras['target_fc4b']),
-            nn.Sigmoid(),
-            TargetFC(paras['target_fc5w'], paras['target_fc5b']),
-        )
-
-    def forward(self, x):
-        q = self.l1(x)
-        # q = F.dropout(q)
-        q = self.l2(q)
-        q = self.l3(q)
-        q = self.l4(q).squeeze()
-        return q
-
-
-class TargetFC(nn.Module):
-    """
-    Fully connection operations for target net
-    Note:
-        Weights & biases are different for different images in a batch,
-        thus here we use group convolution for calculating images in a batch with individual weights & biases.
-    """
-
-    def __init__(self, weight, bias):
-        super(TargetFC, self).__init__()
-        self.weight = weight
-        self.bias = bias
-
-    def forward(self, input_):
-
-        input_re = input_.view(-1, input_.shape[0] * input_.shape[1], input_.shape[2], input_.shape[3])
-        weight_re = self.weight.view(self.weight.shape[0] * self.weight.shape[1], self.weight.shape[2],
-                                     self.weight.shape[3], self.weight.shape[4])
-        bias_re = self.bias.view(self.bias.shape[0] * self.bias.shape[1])
-        out = F.conv2d(input=input_re, weight=weight_re, bias=bias_re, groups=self.weight.shape[0])
-
-        return out.view(input_.shape[0], self.weight.shape[1], input_.shape[2], input_.shape[3])
